+++ import_export/widgets.py
@@ -815,6 +815,21 @@ class ManyToManyWidget(Widget):
         self.field = field
         super().__init__(**kwargs)
 
+    def get_queryset(self, value, row, *args, **kwargs):
+        """
+        Returns a queryset of all objects for this Model.
+
+        Override this method to limit the pool of objects from which related
+        objects are retrieved.
+
+        :param value: The field's value in the dataset.
+        :param row: The dataset's current row.
+        :param \\*args: Optional args.
+        :param \\**kwargs: Optional kwargs.
+        :return: A QuerySet containing all objects for this model.
+        """
+        return self.model.objects.all()
+
     def clean(self, value, row=None, **kwargs):
         """
         Converts a separated string of values into a QuerySet for ManyToMany
@@ -832,12 +847,13 @@ class ManyToManyWidget(Widget):
         """
         if not value:
             return self.model.objects.none()
+        queryset = self.get_queryset(value, row, **kwargs)
         if isinstance(value, (float, int)):
             ids = [int(value)]
         else:
             ids = value.split(self.separator)
             ids = filter(None, [i.strip() for i in ids])
-        return self.model.objects.filter(**{"%s__in" % self.field: ids})
+        return queryset.filter(**{"%s__in" % self.field: ids})
 
     def render(self, value, **kwargs):
         """